Version 1
This is the first version of my C classification neural network. I
learned all of the math for this in my machine learning class and wanted
to implement it throughout the semester in something other than python and make it visually
interesting to look at. There are still improvements but this is a good
project for the first release after the course has wrapped up.